put copy and transform in different files
