+++ Source/Particles/ParticleCreation/Make.package
@@ -1,4 +1,7 @@
+CEXE_sources += ElementaryProcess.cpp
 CEXE_headers += ElementaryProcess.H
+CEXE_headers += CopyParticle.H
+CEXE_headers += TransformParticle.H
 
 INCLUDE_LOCATIONS += $(WARPX_HOME)/Source/Particles/ParticleCreation/
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Particles/ParticleCreation/
